requirements.txt
PyPDF2==3.0.1
Pillow==10.1.0
numpy==1.24.3
reportlab==4.0.7
pytesseract==0.3.10
language-tool-python==2.7.1
